requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.38.5
jinja2==3.1.4
python-multipart==0.0.9

sqlalchemy==2.0.34
passlib[bcrypt]==1.7.4

# ENV / конфиги
python-dotenv==1.0.1

# Базы
psycopg2-binary==2.9.9   # для PostgreSQL в проде

# Прод-сервер (если будешь деплоить через Docker/хостинг)
gunicorn==21.2.0
